Consolidate error and structure generators (#129)
Consolidates Error and structure pojo generation into a single generator. Also refactors the codegen codebase for cleanup and to use more named parameters.
