+++ codegen/core/src/main/java/software/amazon/smithy/java/codegen/generators/SharedSchemasGenerator.java
@@ -95,9 +95,7 @@ public final class SharedSchemasGenerator
                 );
             }
         }
-        writer.openBlock("static {", "}", () -> {
-            writeDeferred(writer, directive, deferred);
-        });
+        writer.openBlock("static {", "}", () -> writeDeferred(writer, directive, deferred));
         writer.popState();
     }
