+++ codegen/core/src/main/java/software/amazon/smithy/java/codegen/generators/StructureGenerator.java
@@ -5,27 +5,73 @@
 
 package software.amazon.smithy.java.codegen.generators;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+import java.time.Instant;
+import java.time.format.DateTimeFormatter;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+import java.util.Objects;
 import java.util.function.Consumer;
-import software.amazon.smithy.codegen.core.directed.GenerateStructureDirective;
+import software.amazon.smithy.codegen.core.SymbolProvider;
+import software.amazon.smithy.codegen.core.directed.ShapeDirective;
 import software.amazon.smithy.java.codegen.CodeGenerationContext;
+import software.amazon.smithy.java.codegen.CodegenUtils;
 import software.amazon.smithy.java.codegen.JavaCodegenSettings;
+import software.amazon.smithy.java.codegen.SymbolProperties;
 import software.amazon.smithy.java.codegen.sections.ClassSection;
+import software.amazon.smithy.java.codegen.sections.GetterSection;
 import software.amazon.smithy.java.codegen.writer.JavaWriter;
+import software.amazon.smithy.java.runtime.core.schema.ModeledSdkException;
+import software.amazon.smithy.java.runtime.core.schema.PresenceTracker;
 import software.amazon.smithy.java.runtime.core.schema.SerializableStruct;
+import software.amazon.smithy.java.runtime.core.serde.DataStream;
+import software.amazon.smithy.model.Model;
+import software.amazon.smithy.model.node.Node;
+import software.amazon.smithy.model.shapes.BigDecimalShape;
+import software.amazon.smithy.model.shapes.BigIntegerShape;
+import software.amazon.smithy.model.shapes.BlobShape;
+import software.amazon.smithy.model.shapes.BooleanShape;
+import software.amazon.smithy.model.shapes.ByteShape;
+import software.amazon.smithy.model.shapes.DocumentShape;
+import software.amazon.smithy.model.shapes.DoubleShape;
+import software.amazon.smithy.model.shapes.EnumShape;
+import software.amazon.smithy.model.shapes.FloatShape;
+import software.amazon.smithy.model.shapes.IntEnumShape;
+import software.amazon.smithy.model.shapes.IntegerShape;
+import software.amazon.smithy.model.shapes.ListShape;
+import software.amazon.smithy.model.shapes.LongShape;
+import software.amazon.smithy.model.shapes.MapShape;
+import software.amazon.smithy.model.shapes.MemberShape;
+import software.amazon.smithy.model.shapes.ServiceShape;
+import software.amazon.smithy.model.shapes.Shape;
+import software.amazon.smithy.model.shapes.ShapeVisitor;
+import software.amazon.smithy.model.shapes.ShortShape;
+import software.amazon.smithy.model.shapes.StringShape;
+import software.amazon.smithy.model.shapes.StructureShape;
+import software.amazon.smithy.model.shapes.TimestampShape;
+import software.amazon.smithy.model.shapes.UnionShape;
+import software.amazon.smithy.model.traits.DefaultTrait;
+import software.amazon.smithy.model.traits.ErrorTrait;
+import software.amazon.smithy.model.traits.StreamingTrait;
+import software.amazon.smithy.model.traits.TimestampFormatTrait;
 import software.amazon.smithy.utils.SmithyInternalApi;
 
 @SmithyInternalApi
-public final class StructureGenerator
-    implements Consumer<GenerateStructureDirective<CodeGenerationContext, JavaCodegenSettings>> {
+public final class StructureGenerator<T extends ShapeDirective<StructureShape, CodeGenerationContext, JavaCodegenSettings>>
+    implements Consumer<T> {
 
     @Override
-    public void accept(GenerateStructureDirective<CodeGenerationContext, JavaCodegenSettings> directive) {
+    public void accept(T directive) {
         var shape = directive.shape();
         directive.context().writerDelegator().useShapeWriter(shape, writer -> {
             writer.pushState(new ClassSection(shape));
+            writer.putContext("isError", shape.hasTrait(ErrorTrait.class));
             writer.putContext("shape", directive.symbol());
             writer.putContext("serializableStruct", SerializableStruct.class);
-            writer.putContext("id", writer.consumer(w -> w.writeIdString(shape)));
+            writer.putContext("sdkException", ModeledSdkException.class);
+            writer.putContext("id", new IdStringGenerator(writer, shape));
             writer.putContext(
                 "schemas",
                 new SchemaGenerator(
@@ -37,19 +83,16 @@ public final class StructureGenerator
                 )
             );
             writer.putContext("properties", new PropertyGenerator(writer, shape, directive.symbolProvider()));
-            writer.putContext(
-                "constructor",
-                new ConstructorGenerator(writer, shape, directive.symbolProvider(), directive.model())
-            );
+            writer.putContext("constructor", new ConstructorGenerator(writer, shape, directive.symbolProvider()));
             writer.putContext(
                 "getters",
                 new GetterGenerator(writer, shape, directive.symbolProvider(), directive.model())
             );
-            writer.putContext("equals", new EqualsGenerator(writer, directive.shape(), directive.symbolProvider()));
-            writer.putContext("hashCode", new HashCodeGenerator(writer, directive.shape(), directive.symbolProvider()));
-            writer.putContext("toString", writer.consumer(JavaWriter::writeToString));
+            writer.putContext("equals", new EqualsGenerator(writer, shape, directive.symbolProvider()));
+            writer.putContext("hashCode", new HashCodeGenerator(writer, shape, directive.symbolProvider()));
+            writer.putContext("toString", new ToStringGenerator(writer));
             writer.putContext(
-                "memberSerializer",
+                "serializer",
                 new StructureSerializerGenerator(
                     writer,
                     directive.symbolProvider(),
@@ -58,32 +101,20 @@ public final class StructureGenerator
                     directive.service()
                 )
             );
-            writer.putContext("builderGetter", writer.consumer(JavaWriter::writeBuilderGetter));
             writer.putContext(
                 "builder",
-                new BuilderGenerator(writer, directive.symbolProvider(), directive.model(), directive.service(), shape)
+                new StructureBuilderGenerator(
+                    writer,
+                    shape,
+                    directive.symbolProvider(),
+                    directive.model(),
+                    directive.service()
+                )
             );
-            // Register inner templates for builder generator
-            writer.putContext(
-                "builderProperties",
-                new StructureBuilderPropertyGenerator(writer, directive.symbolProvider(), directive.model(), shape)
-            );
-            writer.putContext(
-                "builderSetters",
-                new StructureBuilderSetterGenerator(writer, directive.symbolProvider(), directive.model(), shape)
-            );
-            writer.putContext(
-                "errorCorrection",
-                new ErrorCorrectionGenerator(writer, directive.symbolProvider(), directive.model(), shape)
-            );
-            writer.putContext("buildMethod", new StructureBuilderBuildGenerator(writer, shape));
-            writer.putContext(
-                "toBuilder",
-                new ToBuilderGenerator(writer, shape, directive.symbolProvider(), directive.model())
-            );
+            writer.putContext("toBuilder", new ToBuilderGenerator(writer, shape, directive.symbolProvider()));
             writer.write(
                 """
-                    public final class ${shape:T} implements ${serializableStruct:T} {
+                    public final class ${shape:T} ${^isError}implements ${serializableStruct:T}${/isError}${?isError}extends ${sdkException:T}${/isError} {
                         ${id:C|}
 
                         ${schemas:C|}
@@ -98,19 +129,757 @@ public final class StructureGenerator
 
                         ${equals:C|}
 
-                        ${hashCode:C|}
+                        ${^isError}${hashCode:C|}${/isError}
 
-                        ${memberSerializer:C|}
-
-                        ${builderGetter:C|}
+                        ${serializer:C|}
 
                         ${builder:C|}
 
-                        ${toBuilder:C|}
+                        ${^isError}${toBuilder:C|}${/isError}
                     }
                     """
             );
             writer.popState();
         });
+    }
+
+    private record PropertyGenerator(JavaWriter writer, Shape shape, SymbolProvider symbolProvider) implements
+        Runnable {
+        @Override
+        public void run() {
+            for (var member : shape.members()) {
+                // Error message members do not generate property
+                var memberName = symbolProvider.toMemberName(member);
+                if (shape.hasTrait(ErrorTrait.class) && memberName.equals("message")) {
+                    continue;
+                }
+                writer.pushState();
+                writer.putContext("isNullable", CodegenUtils.isNullableMember(member));
+                writer.write(
+                    "private transient final ${?isNullable}$1B${/isNullable}${^isNullable}$1T${/isNullable} $2L;",
+                    symbolProvider.toSymbol(member),
+                    memberName
+                );
+                writer.popState();
+            }
+        }
+    }
+
+    private record ConstructorGenerator(JavaWriter writer, Shape shape, SymbolProvider symbolProvider) implements
+        Runnable {
+        @Override
+        public void run() {
+            writer.openBlock(
+                "private ${shape:T}(Builder builder) {",
+                "}",
+                () -> {
+                    if (shape.hasTrait(ErrorTrait.class)) {
+                        writer.write("super(ID, builder.message);");
+                    }
+
+                    for (var member : shape.members()) {
+                        var memberName = symbolProvider.toMemberName(member);
+                        // Special case for error builders. Message is passed to
+                        // the super constructor. No initializer is created.
+                        if (shape.hasTrait(ErrorTrait.class) && memberName.equals("message")) {
+                            continue;
+                        }
+                        writer.write("this.$1L = builder.$1L;", memberName);
+                    }
+                }
+            );
+        }
+    }
+
+    private static final class GetterGenerator extends ShapeVisitor.Default<Void> implements Runnable {
+        private final JavaWriter writer;
+        private final Shape shape;
+        private final SymbolProvider symbolProvider;
+        private final Model model;
+        private MemberShape member;
+
+        private GetterGenerator(
+            JavaWriter writer,
+            Shape shape,
+            SymbolProvider symbolProvider,
+            Model model
+        ) {
+            this.writer = writer;
+            this.symbolProvider = symbolProvider;
+            this.model = model;
+            this.shape = shape;
+        }
+
+        @Override
+        public void run() {
+            for (var member : shape.members()) {
+                // Error message members do not generate a getter
+                if (shape.hasTrait(ErrorTrait.class) && symbolProvider.toMemberName(member).equals("message")) {
+                    continue;
+                }
+                writer.pushState();
+                writer.putContext("memberName", symbolProvider.toMemberName(member));
+                writer.putContext("member", symbolProvider.toSymbol(member));
+                writer.putContext("isNullable", CodegenUtils.isNullableMember(member));
+                this.member = member;
+                member.accept(this);
+                writer.popState();
+            }
+        }
+
+        @Override
+        protected Void getDefault(Shape shape) {
+            // If the member is not required then prefer the boxed type
+            writer.pushState(new GetterSection(member));
+            writer.write(
+                """
+                    public ${?isNullable}${member:T}${/isNullable}${^isNullable}${member:B}${/isNullable} ${memberName:L}() {
+                        return ${memberName:L};
+                    }
+                    """
+            );
+            writer.popState();
+            return null;
+        }
+
+        @Override
+        public Void listShape(ListShape shape) {
+            writeCollectionGetter(shape);
+            return null;
+        }
+
+        @Override
+        public Void mapShape(MapShape shape) {
+            writeCollectionGetter(shape);
+            return null;
+        }
+
+        private void writeCollectionGetter(Shape shape) {
+            writer.pushState(new GetterSection(member));
+            var shapeSymbol = symbolProvider.toSymbol(shape);
+            writer.putContext("empty", shapeSymbol.expectProperty(SymbolProperties.COLLECTION_EMPTY_METHOD));
+            writer.putContext("wrapper", shapeSymbol.expectProperty(SymbolProperties.COLLECTION_IMMUTABLE_WRAPPER));
+            writer.putContext("collections", Collections.class);
+            writer.write(
+                """
+                    public ${member:T} ${memberName:L}() {${?isNullable}
+                        if (${memberName:L} == null) {
+                            return ${collections:T}.${empty:L};
+                        }${/isNullable}
+                        return ${collections:T}.${wrapper:L}(${memberName:L});
+                    }
+                    """
+            );
+
+            // Write has-er to allow users to check if a given collection was set. Required collections must be set.
+            writer.write(
+                """
+                    public boolean has${memberName:U}() {
+                        return ${?isNullable}${memberName:L} != null${/isNullable}${^isNullable}true${/isNullable};
+                    }
+                    """
+            );
+            writer.popState();
+        }
+
+        @Override
+        public Void memberShape(MemberShape shape) {
+            return model.expectShape(shape.getTarget()).accept(this);
+        }
+    }
+
+    private record EqualsGenerator(JavaWriter writer, Shape shape, SymbolProvider symbolProvider) implements Runnable {
+        @Override
+        public void run() {
+            if (shape.hasTrait(ErrorTrait.class)) {
+                // Errors do not generate equals method
+                return;
+            }
+            writer.write(
+                """
+                    @Override
+                    public boolean equals($T other) {
+                        if (other == this) {
+                            return true;
+                        }
+                        ${C|}
+                    }
+                    """,
+                Object.class,
+                (Runnable) this::writeMemberEquals
+            );
+        }
+
+        private void writeMemberEquals() {
+            // If there are no properties to compare, and they are the same non-null
+            // type then classes should be considered equal, and we can simplify the return
+            if (shape.members().isEmpty()) {
+                writer.writeInlineWithNoFormatting("return other != null && getClass() == other.getClass();");
+                return;
+            }
+            writer.write(
+                """
+                    if (other == null || getClass() != other.getClass()) {
+                        return false;
+                    }
+                    $1T that = ($1T) other;
+                    return ${2C|};""",
+                symbolProvider.toSymbol(shape),
+                (Runnable) this::writePropertyEqualityChecks
+            );
+        }
+
+        private void writePropertyEqualityChecks() {
+            var iter = shape.members().iterator();
+            while (iter.hasNext()) {
+                var member = iter.next();
+                var memberSymbol = symbolProvider.toSymbol(member);
+                writer.pushState();
+                writer.putContext("memberName", symbolProvider.toMemberName(member));
+                // Use `==` instead of `equals` for unboxed primitives
+                if (memberSymbol.expectProperty(SymbolProperties.IS_PRIMITIVE) && !CodegenUtils.isNullableMember(
+                    member
+                )) {
+                    writer.writeInline("${memberName:L} == that.${memberName:L}");
+                } else {
+                    Class<?> comparator = CodegenUtils.isJavaArray(memberSymbol) ? Arrays.class : Objects.class;
+                    writer.writeInline("$T.equals(${memberName:L}, that.${memberName:L})", comparator);
+                }
+                if (iter.hasNext()) {
+                    writer.writeInlineWithNoFormatting(writer.getNewline() + "&& ");
+                }
+                writer.popState();
+            }
+        }
+    }
+
+    private record HashCodeGenerator(JavaWriter writer, Shape shape, SymbolProvider symbolProvider) implements
+        Runnable {
+
+        @Override
+        public void run() {
+            if (shape.hasTrait(ErrorTrait.class)) {
+                // Errors do not generate hashcode
+                return;
+            }
+
+            writer.write(
+                """
+                    @Override
+                    public int hashCode() {
+                        ${C|}
+                    }
+                    """,
+                (Runnable) this::generate
+            );
+        }
+
+        private void generate() {
+            List<String> arrayMemberNames = shape.members()
+                .stream()
+                .filter(member -> CodegenUtils.isJavaArray(symbolProvider.toSymbol(member)))
+                .map(symbolProvider::toMemberName)
+                .toList();
+            List<String> objectMemberNames = shape.members()
+                .stream()
+                .map(symbolProvider::toMemberName)
+                .filter(name -> !arrayMemberNames.contains(name))
+                .toList();
+            writer.pushState();
+            writer.putContext("arr", arrayMemberNames);
+            writer.putContext("obj", objectMemberNames);
+            writer.putContext("objects", Objects.class);
+            if (arrayMemberNames.isEmpty()) {
+                writer.write("return ${objects:T}.hash(${#obj}${value:L}${^key.last}, ${/key.last}${/obj});");
+            } else {
+                writer.putContext("arrays", Arrays.class);
+                writer.write(
+                    """
+                        int result = ${objects:T}.hash(${#obj}${value:L}${^key.last}, ${/key.last}${/obj});
+                        result = 31 * result${#arr} + ${arrays:T}.hashCode(${value:L})${/arr};
+                        return result;
+                        """
+                );
+            }
+            writer.popState();
+        }
+    }
+
+    private record ToBuilderGenerator(JavaWriter writer, StructureShape shape, SymbolProvider symbolProvider) implements
+        Runnable {
+        @Override
+        public void run() {
+            writer.pushState();
+            writer.putContext("members", shape.members().stream().map(symbolProvider::toMemberName).toList());
+            writer.write(
+                """
+                    public Builder toBuilder() {
+                        var builder =  new Builder();
+                        ${#members}builder.${value:L}(this.${value:L});
+                        ${/members}return builder;
+                    }
+                    """
+            );
+            writer.popState();
+        }
+    }
+
+
+    private static final class StructureBuilderGenerator extends BuilderGenerator {
+
+        StructureBuilderGenerator(
+            JavaWriter writer,
+            Shape shape,
+            SymbolProvider symbolProvider,
+            Model model,
+            ServiceShape service
+        ) {
+            super(writer, shape, symbolProvider, model, service);
+        }
+
+        @Override
+        protected void generateErrorCorrection(JavaWriter writer) {
+            if (shape.members().stream().noneMatch(CodegenUtils::isRequiredWithNoDefault)) {
+                // Do not generate error correction if there are no required members with no default.
+                return;
+            }
+
+            writer.write(
+                """
+                    @Override
+                    public ${sdkShapeBuilder:T}<${shape:T}> errorCorrection() {
+                        if (tracker.allSet()) {
+                            return this;
+                        }
+                        ${C|}
+                        return this;
+                    }
+                    """,
+                writer.consumer(this::writeErrorCorrectionMembers)
+            );
+        }
+
+        private void writeErrorCorrectionMembers(JavaWriter writer) {
+            var visitor = new ErrorCorrectionVisitor(writer, symbolProvider, model);
+            for (var member : shape.members()) {
+                if (CodegenUtils.isRequiredWithNoDefault(member)) {
+                    var memberName = symbolProvider.toMemberName(member);
+                    var schemaName = CodegenUtils.toMemberSchemaName(memberName);
+                    visitor.memberShape = member;
+                    writer.openBlock("if (!tracker.checkMember($L)) {", "}", schemaName, () -> {
+                        if (CodegenUtils.hasBuiltinDefault(symbolProvider, model, member)) {
+                            writer.write("tracker.setMember($1L);", schemaName);
+                        } else {
+                            writer.write("$L($C);", memberName, visitor);
+                        }
+                    });
+                }
+            }
+        }
+
+
+        private static final class ErrorCorrectionVisitor extends ShapeVisitor.Default<Void> implements Runnable {
+            private final JavaWriter writer;
+            private final SymbolProvider symbolProvider;
+            private final Model model;
+            private MemberShape memberShape;
+
+            private ErrorCorrectionVisitor(JavaWriter writer, SymbolProvider symbolProvider, Model model) {
+                this.writer = writer;
+                this.symbolProvider = symbolProvider;
+                this.model = model;
+            }
+
+            @Override
+            public void run() {
+                memberShape.accept(this);
+            }
+
+            @Override
+            protected Void getDefault(Shape shape) {
+                throw new IllegalArgumentException("Could not generate error correction value for " + shape);
+            }
+
+            @Override
+            public Void blobShape(BlobShape blobShape) {
+                if (blobShape.hasTrait(StreamingTrait.class)) {
+                    writer.writeInline("$T.ofEmpty()", DataStream.class);
+                } else {
+                    writer.writeInlineWithNoFormatting("new byte[0]");
+                }
+                return null;
+            }
+
+            @Override
+            public Void listShape(ListShape listShape) {
+                writer.writeInline(
+                    "$T.$L",
+                    Collections.class,
+                    symbolProvider.toSymbol(listShape).expectProperty(SymbolProperties.COLLECTION_EMPTY_METHOD)
+                );
+                return null;
+            }
+
+            @Override
+            public Void mapShape(MapShape mapShape) {
+                writer.writeInline(
+                    "$T.$L",
+                    Collections.class,
+                    symbolProvider.toSymbol(mapShape).expectProperty(SymbolProperties.COLLECTION_EMPTY_METHOD)
+                );
+                return null;
+            }
+
+            @Override
+            public Void documentShape(DocumentShape documentShape) {
+                writer.writeInline("null");
+                return null;
+            }
+
+            @Override
+            public Void bigIntegerShape(BigIntegerShape bigIntegerShape) {
+                writer.writeInline("$T.ZERO", BigInteger.class);
+                return null;
+            }
+
+            @Override
+            public Void bigDecimalShape(BigDecimalShape bigDecimalShape) {
+                writer.writeInline("$T.ZERO", BigDecimal.class);
+                return null;
+            }
+
+            @Override
+            public Void stringShape(StringShape stringShape) {
+                writer.writeInline("\"\"");
+                return null;
+            }
+
+            @Override
+            public Void structureShape(StructureShape structureShape) {
+                // Attempts to make an empty structure member. This could fail if the nested struct
+                // has required members.
+                writer.writeInline("$T.builder().build()", symbolProvider.toSymbol(structureShape));
+                return null;
+            }
+
+            @Override
+            public Void unionShape(UnionShape unionShape) {
+                // TODO: implement for unions
+                writer.writeInline("null");
+                return null;
+            }
+
+
+            @Override
+            public Void enumShape(EnumShape shape) {
+                // TODO: Implement
+                return null;
+            }
+
+            @Override
+            public Void intEnumShape(IntEnumShape shape) {
+                // TODO: Implement
+                return null;
+            }
+
+            @Override
+            public Void timestampShape(TimestampShape timestampShape) {
+                writer.write("$T.EPOCH", Instant.class);
+                return null;
+            }
+
+            @Override
+            public Void memberShape(MemberShape memberShape) {
+                return model.expectShape(memberShape.getTarget()).accept(this);
+            }
+        }
+
+        @Override
+        protected void generateProperties(JavaWriter writer) {
+            writer.pushState();
+
+            // Add any static default properties
+            var defaultVisitor = new DefaultInitializerGenerator(writer, model, symbolProvider);
+            for (var member : shape.members()) {
+                if (member.hasNonNullDefault()
+                    && symbolProvider.toSymbol(member).expectProperty(SymbolProperties.REQUIRES_STATIC_DEFAULT)
+                ) {
+                    defaultVisitor.member = member;
+                    writer.write(
+                        "private static final $T $L = $C;",
+                        symbolProvider.toSymbol(member),
+                        CodegenUtils.toDefaultValueName(symbolProvider.toMemberName(member)),
+                        defaultVisitor
+                    );
+                }
+            }
+
+            // Add presence tracker
+            writer.putContext("tracker", PresenceTracker.class);
+            writer.write("private final ${tracker:T} tracker = ${tracker:T}.of(SCHEMA);");
+
+            // Add non-static builder properties
+            for (var member : shape.members()) {
+                var memberName = symbolProvider.toMemberName(member);
+                if (CodegenUtils.isStreamingBlob(model.expectShape(member.getTarget()))) {
+                    // Streaming blobs need a custom initializer
+                    writer.write(
+                        "private $1T $2L = $1T.ofEmpty();",
+                        DataStream.class,
+                        memberName
+                    );
+                    continue;
+                }
+                writer.pushState();
+                writer.putContext("nullable", CodegenUtils.isNullableMember(member));
+                writer.putContext("default", member.hasNonNullDefault());
+                writer.putContext(
+                    "static",
+                    symbolProvider.toSymbol(member).expectProperty(SymbolProperties.REQUIRES_STATIC_DEFAULT)
+                );
+                defaultVisitor.member = member;
+                writer.write(
+                    "private ${^nullable}$1T${/nullable}${?nullable}$1B${/nullable} $2L${?default} = ${?static}$3L${/static}${^static}$4C${/static}${/default};",
+                    symbolProvider.toSymbol(member),
+                    memberName,
+                    CodegenUtils.toDefaultValueName(memberName),
+                    defaultVisitor
+                );
+                writer.popState();
+            }
+            writer.popState();
+        }
+
+        @Override
+        protected void generateSetters(JavaWriter writer) {
+            writer.pushState();
+            writer.putContext("objects", Objects.class);
+            for (var member : shape.members()) {
+                writer.pushState();
+                writer.putContext("memberName", symbolProvider.toMemberName(member));
+                writer.putContext("memberSymbol", symbolProvider.toSymbol(member));
+                writer.putContext("tracked", CodegenUtils.isRequiredWithNoDefault(member));
+                writer.putContext("check", CodegenUtils.requiresSetterNullCheck(symbolProvider, member));
+                writer.putContext("schemaName", CodegenUtils.toMemberSchemaName(symbolProvider.toMemberName(member)));
+
+                // If streaming blob then a setter must be added to allow
+                // operation to set on builder.
+                if (CodegenUtils.isStreamingBlob(model.expectShape(member.getTarget()))) {
+                    writer.write("""
+                        @Override
+                        public void setDataStream($T stream) {
+                            ${memberName:L}(stream);
+                        }
+                        """, DataStream.class);
+                }
+                writer.write(
+                    """
+                        public Builder ${memberName:L}(${memberSymbol:T} ${memberName:L}) {
+                            this.${memberName:L} = ${?check}${objects:T}.requireNonNull(${/check}${memberName:L}${?check}, "${memberName:L} cannot be null")${/check};${?tracked}
+                            tracker.setMember(${schemaName:L});${/tracked}
+                            return this;
+                        }
+                        """
+                );
+                writer.popState();
+            }
+            writer.popState();
+        }
+
+        @Override
+        protected void generateBuild(JavaWriter writer) {
+            writer.pushState();
+            writer.putContext(
+                "hasRequiredMembers",
+                shape.members().stream().anyMatch(CodegenUtils::isRequiredWithNoDefault)
+            );
+            writer.write("""
+                @Override
+                public ${shape:T} build() {${?hasRequiredMembers}
+                    tracker.validate();${/hasRequiredMembers}
+                    return new ${shape:T}(this);
+                }
+                """);
+            writer.popState();
+        }
+
+        /**
+         * Adds default values to builder properties.
+         */
+        private static final class DefaultInitializerGenerator extends ShapeVisitor.DataShapeVisitor<Void> implements
+            Runnable {
+            private final JavaWriter writer;
+            private final Model model;
+            private final SymbolProvider symbolProvider;
+            private MemberShape member;
+            private Node defaultValue;
+
+            DefaultInitializerGenerator(
+                JavaWriter writer,
+                Model model,
+                SymbolProvider symbolProvider
+            ) {
+                this.writer = writer;
+                this.model = model;
+                this.symbolProvider = symbolProvider;
+            }
+
+            @Override
+            public void run() {
+                if (member.hasNonNullDefault()) {
+                    this.defaultValue = member.expectTrait(DefaultTrait.class).toNode();
+                    member.accept(this);
+                }
+            }
+
+            @Override
+            public Void blobShape(BlobShape blobShape) {
+                throw new UnsupportedOperationException("Blob default value cannot be set.");
+            }
+
+            @Override
+            public Void booleanShape(BooleanShape booleanShape) {
+                writer.write("$L", defaultValue.expectBooleanNode().getValue());
+                return null;
+            }
+
+            @Override
+            public Void listShape(ListShape listShape) {
+                // Note that Lists can _ONLY_ have empty maps as the default,
+                // so we do not need to check the default value. See:
+                // https://github.com/smithy-lang/smithy/blob/main/designs/defaults-and-model-evolution.md
+                writer.write(
+                    "$T.$L",
+                    Collections.class,
+                    symbolProvider.toSymbol(listShape).expectProperty(SymbolProperties.COLLECTION_EMPTY_METHOD)
+                );
+                return null;
+            }
+
+            @Override
+            public Void mapShape(MapShape mapShape) {
+                // Note that Maps can _ONLY_ have empty maps as the default,
+                // so we do not need to check the default value. See:
+                // https://github.com/smithy-lang/smithy/blob/main/designs/defaults-and-model-evolution.md
+                writer.write(
+                    "$T.$L",
+                    Collections.class,
+                    symbolProvider.toSymbol(mapShape).expectProperty(SymbolProperties.COLLECTION_EMPTY_METHOD)
+                );
+                return null;
+            }
+
+            @Override
+            public Void byteShape(ByteShape byteShape) {
+                // Bytes duplicate the integer toString method
+                writer.write("$L", defaultValue.expectNumberNode().getValue().intValue());
+                return null;
+            }
+
+            @Override
+            public Void shortShape(ShortShape shortShape) {
+                // Shorts duplicate the int toString method
+                writer.write("$L", defaultValue.expectNumberNode().getValue().intValue());
+                return null;
+            }
+
+            @Override
+            public Void integerShape(IntegerShape integerShape) {
+                writer.write("$L", defaultValue.expectNumberNode().getValue().intValue());
+                return null;
+            }
+
+            @Override
+            public Void longShape(LongShape longShape) {
+                writer.write("$LL", defaultValue.expectNumberNode().getValue().longValue());
+                return null;
+            }
+
+            @Override
+            public Void floatShape(FloatShape floatShape) {
+                writer.write("$Lf", defaultValue.expectNumberNode().getValue().floatValue());
+                return null;
+            }
+
+            @Override
+            public Void documentShape(DocumentShape documentShape) {
+                throw new UnsupportedOperationException("Document shape defaults cannot be set.");
+            }
+
+            @Override
+            public Void doubleShape(DoubleShape doubleShape) {
+                writer.write("$L", defaultValue.expectNumberNode().getValue().doubleValue());
+                return null;
+            }
+
+            @Override
+            public Void bigIntegerShape(BigIntegerShape bigIntegerShape) {
+                writer.write("$T.valueOf($L)", BigInteger.class, defaultValue.expectNumberNode().getValue().intValue());
+                return null;
+            }
+
+            @Override
+            public Void bigDecimalShape(BigDecimalShape bigDecimalShape) {
+                writer.write(
+                    "$T.valueOf($L)",
+                    BigDecimal.class,
+                    defaultValue.expectNumberNode().getValue().doubleValue()
+                );
+                return null;
+            }
+
+            @Override
+            public Void stringShape(StringShape stringShape) {
+                writer.write("$S", defaultValue.expectStringNode().getValue());
+                return null;
+            }
+
+            @Override
+            public Void structureShape(StructureShape structureShape) {
+                throw new UnsupportedOperationException("Structure shape defaults cannot be set.");
+            }
+
+            @Override
+            public Void unionShape(UnionShape unionShape) {
+                throw new UnsupportedOperationException("Union shape defaults cannot be set.");
+
+            }
+
+            @Override
+            public Void memberShape(MemberShape memberShape) {
+                return model.expectShape(memberShape.getTarget()).accept(this);
+            }
+
+            @Override
+            public Void timestampShape(TimestampShape timestampShape) {
+                Instant value;
+                if (member.hasTrait(TimestampFormatTrait.class)) {
+                    value = switch (member.expectTrait(TimestampFormatTrait.class).getFormat()) {
+                        case EPOCH_SECONDS -> Instant.ofEpochMilli(
+                            defaultValue.expectNumberNode().getValue().longValue()
+                        );
+                        case HTTP_DATE -> Instant.from(
+                            DateTimeFormatter.RFC_1123_DATE_TIME.parse(defaultValue.expectStringNode().getValue())
+                        );
+                        default -> instantFromDefaultTimestamp(defaultValue);
+                    };
+                } else {
+                    value = instantFromDefaultTimestamp(defaultValue);
+                }
+                writer.write("$T.ofEpochMilli($LL)", Instant.class, value.toEpochMilli());
+                return null;
+            }
+
+            private static Instant instantFromDefaultTimestamp(Node defaultValue) {
+                if (defaultValue.isNumberNode()) {
+                    return Instant.ofEpochSecond(defaultValue.expectNumberNode().getValue().longValue());
+                } else if (defaultValue.isStringNode()) {
+                    return Instant.parse(defaultValue.expectStringNode().getValue());
+                }
+                throw new IllegalArgumentException(
+                    "Invalid timestamp value node: " + defaultValue + "Expected string or number"
+                );
+            }
+        }
     }
 }
