+++ codegen/core/src/test/java/software/amazon/smithy/java/codegen/utils/TestJavaCodegen.java
@@ -22,7 +22,6 @@ import software.amazon.smithy.java.codegen.CodeGenerationContext;
 import software.amazon.smithy.java.codegen.JavaCodegenIntegration;
 import software.amazon.smithy.java.codegen.JavaCodegenSettings;
 import software.amazon.smithy.java.codegen.JavaSymbolProvider;
-import software.amazon.smithy.java.codegen.generators.ExceptionGenerator;
 import software.amazon.smithy.java.codegen.generators.ListGenerator;
 import software.amazon.smithy.java.codegen.generators.MapGenerator;
 import software.amazon.smithy.java.codegen.generators.SharedSchemasGenerator;
@@ -71,13 +70,12 @@ public class TestJavaCodegen implements
 
     @Override
     public void generateStructure(GenerateStructureDirective<CodeGenerationContext, JavaCodegenSettings> directive) {
-        new StructureGenerator().accept(directive);
-
+        new StructureGenerator<>().accept(directive);
     }
 
     @Override
     public void generateError(GenerateErrorDirective<CodeGenerationContext, JavaCodegenSettings> directive) {
-        new ExceptionGenerator().accept(directive);
+        new StructureGenerator<>().accept(directive);
     }
 
     @Override
